First commit with a helper class, a widget and a behavior.
In the first commit, there are 2 components. One of them is a helper
class - "AHTML" which is inherited from CHTML for creating 'ng' HTML
element. The other one is a widget to inject angular JS script files.
Also there is a AngularJSConfigBehavior which uses to import all
components to project easily.